+++ packages/shopify/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @weaverse/shopify
 
+## 5.8.0
+
+### Patch Changes
+
+- @weaverse/react@5.8.0
+
 ## 5.7.1
 
 ### Patch Changes